Restructured the DB-Model for Groups.
The initial refactoring of the Groups and their relationship to the Users.
Instead of storing a group-ID as interger-field, now the hibernate
mechanism of a many-to-many relationship is used to model the user-groups and
their corresponding users.
In this commit the database structure (models) is adapted and the broken
direct dependecies are fixed.
The DB-init-script is also modified. The anonymousUser is now added to the
default-group.