New annotation to set access-restrictions on Hibernate-models.
Introduced an annotation to restrict the access of Hibernate model
instances, which represent database tables. The annotation can be set on
the class itself and should be parameterized with a value offered by the
enum RestrictionType, e.g. 'GROUP' to allow only group-related access.
A parser for accessing the annotation and its parameter during runtime is
also implemented.