Remove automatic group-detection from DB-DAO.
The mechanism of the automatic detection of group-dependency for a model
within the DB-DAO is now removed. As other security mechanisms are done at
service-level, filtering/restrictions by groups should be done at
service-level as well.
Therefore some extensions of the DB-DAO will be needed in the future to
pass custom Hibernate Restrictions and Criterions to certain API-methods.
As a show-case of the strategy a new getUserByName method was implemented
accepting custom Criterion/Restriction.
The broken API-calls, due to the changed API-methods of the DB-DAO are
fixed as well in this commit.