Refactored the web-interface for group-creation.
Restructured the web-interface for the creation of a group entity.
Multiple entity-creation is replaced by single creation. Now it is also allowed
to create a group with the role 'ROLE_ADMIN', which implies that the
logged in user will be the group-leader. The automatic creation of a
group-leader in case od beeing 'SUPERADMIN' has been moved to a private
method.