Refactored web-interface for user-deletion.
Refactored the web-interface which is responsible for the deletion of
user objects in the database. This was necessary beacuse of the changed
user-group relationship (many-to-many). Now all group relationships are
removed before deletion in order to prevent constraint violations.
In case of deleting as admin only users in your own group can be deleted.
Two new DB-DAO methods which are needed for this deletion process are also
included in this commit.