Parse formatted date strings when deserializing.
When deserializing JSON strings sent from clients, we usually accept numeric
timestamps for date columns. With this change we can now also accept strings
in a common date format.

The current implementation will successfully parse dates passed as e.g.
`30.04.2013 13:43:55` because the `java.text.SimpleDateFormat` used is
`dd.MM.yyyy HH:mm:ss`. Future implementations might enhance this approach
with more sophisticated methods to both check whether we are deserializing a
formatted date and more formats which are tried in order to transform a string
to a proper date.