Secure getAppContext only by spring-security-XML
The web service for requesting the app-context has been secured by the
applicationContext-security.xml and by a Java annotation within the
ShogunService class. These two mechanisms were in conflict to each other:
While the XML allowed anonymous-users to request the app-context it was
forbidden by the annotation. So anonymous-users could not request their
own app-context although it was explicitly allowed in the XML. A
consequence was that possible web-open applications could not been
started. This change removes the annotaion and the web-service is
explicitly secured by the applicationContext-security.xml.